+++ silzila-frontend/src/Components/Charts/CrossTab/CrossTab.css
@@ -67,6 +67,21 @@
 	text-align: left;
 }
 
+.TableHeader{
+	border: 1px solid rgb(222, 222, 222);
+	cursor: default;
+	font-weight: bold;
+	font-size: 12px;
+	white-space: nowrap;
+	padding-left: 8px;
+	padding-right: 4px;
+	text-align: center;
+	text-overflow: ellipsis;
+	left: 0;
+	top: auto;
+	text-align: left;
+}
+
 .CrossTabRow {
 	line-height: 1.2em;
 }
