+++ silzila-frontend/src/Components/GraphArea/GraphArea.tsx
@@ -38,7 +38,7 @@ import OpenInFullIcon from "@mui/icons-material/OpenInFull";
 import TextEditor from "../Charts/TextEditor";
 import CrossTabChart from "../Charts/CrossTab/CrossTabChart";
 import FetchData from "../ServerCall/FetchData";
-import { getChartData } from "../ChartAxes/ChartAxes";
+import { getChartData } from "../ChartAxes/ChartData";
 import {
 	updateChartMargins,
 	updateQueryResult,
@@ -553,6 +553,7 @@ const GraphArea = ({
 			chartGroup,
 			dashBoardGroup,
 			propKey,
+			"Chartaxes",
 			token,
 			true
 		).then(async data => {
